feat: GetData() 함수의 인수 타입을 int32로 변경한다
- 사유는 굳이 포트폴리오 단계에서 int64로 할 유인이 없기 때문

+++ ServerCore/DBBind.h
@@ -124,7 +124,7 @@ public:
 		_column_flag |= (1LL << index);
 	}
 
-	bool GetId(OUT int64& id)
+	bool GetId(OUT int32& id)
 	{
 		return _db_connection.GetData(id);
 	}
